+++ src/Saucer.cpp
@@ -1,7 +1,7 @@
 #include "../Headers/Saucer.h"
 
 // Saucer is the thing that occasionally goes across the top of the screen for extra points
-Saucer::Saucer(int health_, int movespeed_, int fireRate_, float xpos_, float ypos_, bool collides_, Game* game_)
+Saucer::Saucer(int health_, int movespeed_, int fireRate_, int xpos_, int ypos_, bool collides_, Game* game_)
 	: Entity(health_, movespeed_, fireRate_, xpos_, ypos_, collides_, game_) {
 	game = game_;
 	health = health_;
