+++ src/TextManager.cpp
@@ -4,7 +4,7 @@ TextManager::TextManager(FontManager* fontManager_) {
 	fontManager = fontManager_;
 }
 
-void TextManager::createText(const std::string& name, const std::string& fontName, int size, sf::Color color, std::string string, float xpos, float ypos) {
+void TextManager::createText(const std::string& name, const std::string& fontName, int size, sf::Color color, std::string string, int xpos, int ypos) {
 	sf::Text text;
 	sf::Font* font = &fontManager->getFontRef(fontName);
 
@@ -36,7 +36,7 @@ void TextManager::updateTextSize(const std::string& textName, int size) {
 	texts[textName] = text;
 }
 
-void TextManager::updateTextPosition(const std::string& textName, float xpos, float ypos) {
+void TextManager::updateTextPosition(const std::string& textName, int xpos, int ypos) {
 	sf::Text text = getTextRef(textName);
 	text.setPosition(xpos, ypos);
 	texts[textName] = text;
